Fix Python reference count leak.
Callbacks set by Persistent::SetWeak aren't guaranteed to be called by
V8, even if the context gets destroyed. As such, the majority of Python
objects passed into the context will never be freed.

To fix this, we replace js_object_cache (which was a
weakref.WeakKeyDictionary) with a normal dict. This ensures the
reference count is managed with the lifetime of the dictionary.